LPS-137925 Retrieve Quote adjustments
