+++ modules/apps/site-initializer/site-initializer-raylife/extra/remote-app/src/common/services/liferay/adapter.js
@@ -23,7 +23,7 @@ const adaptToBusinessType = (data) =>
 const adaptToRaylifeApplicationToForm = (data) => {
 	const basics = {
 		applicationId: data.id,
-		businessCategoryId: '43004',
+		businessCategoryId: data.businessCategoryId,
 		businessInformation: {
 			business: {
 				email: data.email,
@@ -40,12 +40,12 @@ const adaptToRaylifeApplicationToForm = (data) => {
 			firstName: data.firstName,
 			lastName: data.lastName,
 		},
-		businessSearch: 'pet',
-		product: 'Pet Stores',
-		productQuote: 43359,
+		businessSearch: data.businessSearch,
+		product: data.product,
+		productQuote: Number(data.productQuote),
 		properties: {
-			naics: '453910',
-			segment: 'Retail',
+			naics: data.naics,
+			segment: data.segment,
 		},
 	};
 
@@ -110,7 +110,9 @@ const adaptToFormApplicationRequest = (form) => ({
 	annualPayrollForEmployees: form?.employees?.annualPayrollForEmployees,
 	annualPayrollForOwner: form?.employees?.annualPayrollForOwner,
 	buildingSquareFeetOccupied: form?.property?.buildingSquareFeetOccupied,
+	businessCategoryId: form?.basics?.businessCategoryId,
 	businessOperatesYearRound: form?.employees?.businessOperatesYearRound,
+	businessSearch: form?.basics?.businessSearch,
 	city: form?.basics?.businessInformation?.business?.location?.city,
 	doOwnBuildingAtAddress: form?.property?.doOwnBuildingAtAddress,
 	email: form?.basics?.businessInformation?.business?.email,
@@ -126,10 +128,14 @@ const adaptToFormApplicationRequest = (form) => ({
 	isThereSwimming: form?.property?.isThereSwimming,
 	lastName: form?.basics?.businessInformation?.lastName,
 	legalEntity: form?.business?.legalEntity,
+	naics: form?.basics?.properties?.naics,
 	overallSales: form?.business?.overallSales,
 	partTimeEmployees: form?.employees?.partTimeEmployees,
 	phone: form?.basics?.businessInformation?.business?.phone,
+	product: form?.basics?.product,
+	productQuote: `${form?.basics?.productQuote}`,
 	salesMerchandise: form?.business?.salesMerchandise,
+	segment: form?.basics?.properties?.segment,
 	startBusinessAtYear: form?.employees?.startBusinessAtYear,
 	state: form?.basics?.businessInformation?.business?.location?.state,
 	stories: form?.property?.stories,
