+++ modules/apps/site-initializer/site-initializer-raylife/extra/remote-app/src/routes/get-a-quote/components/containers/Forms/Basics/BusinessInformation/index.js
@@ -7,8 +7,10 @@ import {WebsiteControlledInput} from '~/common/components/connectors/Controlled/
 import {PhoneControlledInput} from '~/common/components/connectors/Controlled/Input/WithMask/Phone';
 import {CardFormActionsWithSave} from '~/common/components/fragments/Card/FormActionsWithSave';
 import {useCustomEvent} from '~/common/hooks/useCustomEvent';
+import {STORAGE_KEYS, Storage} from '~/common/services/liferay/storage';
 import {TIP_EVENT} from '~/common/utils/events';
 import useFormActions from '~/routes/get-a-quote/hooks/useFormActions';
+import {useProductQuotes} from '~/routes/get-a-quote/hooks/useProductQuotes';
 import {useStepWizard} from '~/routes/get-a-quote/hooks/useStepWizard';
 import {AVAILABLE_STEPS} from '~/routes/get-a-quote/utils/constants';
 
@@ -16,8 +18,17 @@ import {BusinessInformationAddress} from './Address';
 
 const setFormPath = (value) => `basics.businessInformation.${value}`;
 
+const getSelectedProductName = () => {
+	try {
+		return JSON.parse(Storage.getItem(STORAGE_KEYS.PRODUCT))?.productName;
+	} catch (error) {
+		return '';
+	}
+};
+
 export function FormBasicBusinessInformation({form}) {
 	const {selectedStep} = useStepWizard();
+	const {productQuotes} = useProductQuotes();
 	const [dispatchEvent] = useCustomEvent(TIP_EVENT);
 	const {onNext, onPrevious, onSave} = useFormActions(
 		form,
@@ -29,6 +40,7 @@ export function FormBasicBusinessInformation({form}) {
 	const {
 		control,
 		formState: {isValid},
+		setValue,
 	} = useFormContext();
 
 	const onFirstNameSettled = () => {
@@ -58,6 +70,18 @@ export function FormBasicBusinessInformation({form}) {
 			templateName: 'hi-template',
 		});
 	}, []);
+
+	const defaultProductId = productQuotes.find(
+		({title}) => title === getSelectedProductName()
+	)?.id;
+
+	useEffect(() => {
+		if (defaultProductId && !form.basics.productQuote) {
+			setValue('basics.productQuote', defaultProductId);
+		}
+
+		// eslint-disable-next-line react-hooks/exhaustive-deps
+	}, [defaultProductId, form.basics.productQuote]);
 
 	return (
 		<div className="card">
