+++ modules/apps/site-initializer/site-initializer-raylife/extra/remote-app/src/routes/get-a-quote/components/containers/Forms/Basics/ProductQuote.js
@@ -1,9 +1,8 @@
-import React, {useEffect} from 'react';
+import React from 'react';
 import {Controller, useFormContext} from 'react-hook-form';
 import {MoreInfoButton} from '~/common/components/fragments/Buttons/MoreInfo';
 import {CardFormActionsWithSave} from '~/common/components/fragments/Card/FormActionsWithSave';
 import {Radio} from '~/common/components/fragments/Forms/Radio';
-import {STORAGE_KEYS, Storage} from '~/common/services/liferay/storage';
 import {TIP_EVENT} from '~/common/utils/events';
 import useFormActions from '~/routes/get-a-quote/hooks/useFormActions';
 import {useProductQuotes} from '~/routes/get-a-quote/hooks/useProductQuotes';
@@ -11,17 +10,8 @@ import {useStepWizard} from '~/routes/get-a-quote/hooks/useStepWizard';
 import {useTriggerContext} from '~/routes/get-a-quote/hooks/useTriggerContext';
 import {AVAILABLE_STEPS} from '~/routes/get-a-quote/utils/constants';
 
-const getSelectedProductName = () => {
-	try {
-		return JSON.parse(Storage.getItem(STORAGE_KEYS.PRODUCT))?.productName;
-	}
-	catch (error) {
-		return '';
-	}
-};
-
 export function FormBasicProductQuote({form}) {
-	const {control, setValue} = useFormContext();
+	const {control} = useFormContext();
 	const {selectedStep} = useStepWizard();
 	const {productQuotes} = useProductQuotes();
 	const {onNext, onPrevious, onSave} = useFormActions(
@@ -30,18 +20,7 @@ export function FormBasicProductQuote({form}) {
 		AVAILABLE_STEPS.BUSINESS
 	);
 
-	const defaultProductId = productQuotes.find(
-		({title}) => title === getSelectedProductName()
-	)?.id;
-
 	const {isSelected, updateState} = useTriggerContext();
-
-	useEffect(() => {
-		if (defaultProductId && !form.basics.productQuote) {
-			setValue('basics.productQuote', defaultProductId);
-		}
-		// eslint-disable-next-line react-hooks/exhaustive-deps
-	}, [defaultProductId, form.basics.productQuote]);
 
 	return (
 		<div className="card">
@@ -52,7 +31,7 @@ export function FormBasicProductQuote({form}) {
 					<fieldset className="content-column" id="productQuote">
 						<Controller
 							control={control}
-							defaultValue={defaultProductId}
+							defaultValue={form?.basics?.productQuote}
 							name="basics.productQuote"
 							render={({field}) =>
 								productQuotes.map((quote) => (
