+++ modules/apps/site-initializer/site-initializer-raylife/src/main/resources/site-initializer/object-definitions/raylife-application.json
@@ -84,6 +84,17 @@
 			"required": false,
 			"type": "String"
 		},
+
+		{
+			"indexed": true,
+			"indexedAsKeyword": false,
+			"label": {
+				"en_US": "Business Search"
+			},
+			"name": "businessSearch",
+			"required": false,
+			"type": "String"
+		},
 		{
 			"indexed": true,
 			"indexedAsKeyword": false,
@@ -238,6 +249,16 @@
 			"indexed": true,
 			"indexedAsKeyword": false,
 			"label": {
+				"en_US": "Naics"
+			},
+			"name": "naics",
+			"required": false,
+			"type": "String"
+		},
+		{
+			"indexed": true,
+			"indexedAsKeyword": false,
+			"label": {
 				"en_US": "Overall Sales"
 			},
 			"name": "overallSales",
@@ -268,6 +289,16 @@
 			"indexed": true,
 			"indexedAsKeyword": false,
 			"label": {
+				"en_US": "Product"
+			},
+			"name": "product",
+			"required": false,
+			"type": "String"
+		},
+		{
+			"indexed": true,
+			"indexedAsKeyword": false,
+			"label": {
 				"en_US": "Product Quote"
 			},
 			"name": "productQuote",
@@ -281,6 +312,16 @@
 				"en_US": "Sales Merchandise"
 			},
 			"name": "salesMerchandise",
+			"required": false,
+			"type": "String"
+		},
+		{
+			"indexed": true,
+			"indexedAsKeyword": false,
+			"label": {
+				"en_US": "Segment"
+			},
+			"name": "segment",
 			"required": false,
 			"type": "String"
 		},
